renamed players list
